chore: Move Hasura to lib folder
