+++ api.planx.uk/modules/webhooks/service/paymentRequestEvents/index.test.ts
@@ -1,11 +1,11 @@
 import supertest from "supertest";
 import app from "../../../../server";
-import { createScheduledEvent } from "../../../../hasura/metadata";
+import { createScheduledEvent } from "../../../../lib/hasura/metadata";
 import { CreatePaymentEvent } from "./schema";
 
 const { post } = supertest(app);
 
-jest.mock("../../../../hasura/metadata");
+jest.mock("../../../../lib/hasura/metadata");
 const mockedCreateScheduledEvent = createScheduledEvent as jest.MockedFunction<
   typeof createScheduledEvent
 >;
